+++ src/config.php
@@ -2,6 +2,9 @@
 
 return [
     'enableCertificates' => true,
+    'enableCourseUrls' => false,
+    'courseUriFormat' => 'courses/{slug}',
+    'courseTemplate' => '',
     'certificateTemplatePath' => '',
     'autoEnrollOnPurchase' => true,
     'autoIssueCertificate' => true,